ragMolmind/requirements.txt
fastapi
uvicorn 
pydantic 
langchain-community 
langchain-ollama 
langchain-groq 
langchain 
python-dotenv 
faiss-cpu 
pymupdf 
arxiv